carbid/requirements.txt
# Web Framework and API
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
pydantic[email]==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Database
pymysql==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1
mysqlclient==2.2.0

# Blockchain
web3==6.11.3
eth-account==0.9.0
eth-utils==2.3.0
eth-typing==3.5.0
hexbytes==0.3.1

# Authentication & Security
bcrypt==4.0.1
pyjwt==2.8.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.5

# File Handling & Processing
Pillow==10.1.0
aiofiles==23.2.1
python-magic==0.4.27

# CORS & Middleware
cors==1.0.1

# Utils
setuptools==69.0.3
wheel==0.42.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Logging & Monitoring
logging==0.4.9.6
structlog==23.2.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Optional but recommended for production
gunicorn==21.2.0
supervisor==4.2.5
redis==5.0.1
celery==5.3.4

# Environment & Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# API Documentation
swagger-ui-bundle==0.0.9

# Performance & Caching
ujson==5.8.0
orjson==3.9.10
aiocache==0.12.2

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Time Zone Support
pytz==2023.3.post1

# IPFS Integration
ipfshttpclient==0.8.0
multiaddr==0.0.9

# Data Validation
email-validator==2.1.0.post1
pydantic[email]==2.4.2

# Security Headers
secure==0.3.0

# Background Tasks
APScheduler==3.10.4

# Compression
python-snappy==0.6.1

# Static File Serving
aiofiles==23.2.1
whitenoise==6.6.0